requirements.txt
asgiref==3.2.10
astroid==2.2.5
Click==7.0
dj-database-url==0.5.0
Django~=3.0.7
djangorestframework==3.11.0
djangorestframework_simplejwt==4.5.0
django-cors-headers==3.5.0
dj-database-url==0.5.0
django-heroku==0.3.1
Flask==1.0.2
gunicorn==20.0.4
isort==4.3.13
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
psycopg2==2.8.6
pylint==2.3.
pycurl==7.43.0.5
pytest-cov==2.10.1
pytest-django==4.1.0
pytz==2020.1
six==1.12.0
sqlparse==0.4.1
typed-ast==1.4.1
Werkzeug==0.14.1
whitenoise==5.2.0
wrapt==1.11.1
django-cors-headers==3.5.0
zoomus==1.1.3